+++ src/imio/dashboard/columns.py
@@ -1,7 +1,15 @@
 # encoding: utf-8
 
-from collective.contact.core.content.organization import IOrganization
 from collective.eeafaceted.z3ctable.columns import PrettyLinkColumn
+
+import pkg_resources
+
+HAS_CONTACT_CORE = True
+try:
+    pkg_resources.get_distribution('collective.contact.core')
+    from collective.contact.core.content.organization import IOrganization
+except pkg_resources.DistributionNotFound:
+    HAS_CONTACT_CORE = False
 
 
 class ContactPrettyLinkColumn(PrettyLinkColumn):
@@ -23,7 +31,7 @@ class ContactPrettyLinkColumn(PrettyLinkColumn):
 
         current_org_css_class = None
         obj = self._getObject(item)
-        if IOrganization.providedBy(obj):
+        if HAS_CONTACT_CORE and IOrganization.providedBy(obj):
             current_org_css_class = '___'.join([org.id for org in obj.get_organizations_chain()])
 
         if current_org_css_class: